chore: remove circular deps
