+++ src/fluent/types.ts
@@ -5,15 +5,20 @@ import { KubernetesListObject, KubernetesObject } from "@kubernetes/client-node"
 import { Operation } from "fast-json-patch";
 import type { PartialDeep } from "type-fest";
 import { RequestInit } from "undici";
-import { GenericClass, GroupVersionKind } from "../types";
-import { WatchCfg, Watcher } from "./watch";
+import { GenericClass } from "../types";
+import { WatchCfg, WatchAction } from "../types/shared";
 import https from "https";
 import { SecureClientSessionOptions } from "http2";
+import { EventEmitter } from "stream";
 
 /*
  * Watch Class Type
  */
-export type WatcherType<T extends GenericClass> = Watcher<T>;
+export interface WatcherType {
+  start(): Promise<AbortController>;
+  close(): void;
+  events: EventEmitter;
+}
 
 /**
  * Fetch options and server URL
@@ -35,17 +40,6 @@ export interface Options {
   agent?: https.Agent & { options?: AgentOptions };
 }
 
-/**
- * The Phase matched when using the K8s Watch API.
- */
-export enum WatchPhase {
-  Added = "ADDED",
-  Modified = "MODIFIED",
-  Deleted = "DELETED",
-  Bookmark = "BOOKMARK",
-  Error = "ERROR",
-}
-
 export type FetchMethods =
   | "APPLY"
   | "DELETE"
@@ -56,14 +50,6 @@ export type FetchMethods =
   | "POST"
   | "PUT"
   | "WATCH";
-
-export interface Filters {
-  kindOverride?: GroupVersionKind;
-  fields?: Record<string, string>;
-  labels?: Record<string, string>;
-  name?: string;
-  namespace?: string;
-}
 
 /**
  * Get the resource or resources matching the filters.
@@ -114,7 +100,7 @@ export type K8sFilteredActions<T extends GenericClass, K extends KubernetesObjec
    * @param watchCfg - (optional) watch configuration
    * @returns a watch controller
    */
-  Watch: (callback: WatchAction<T>, watchCfg?: WatchCfg) => Watcher<T>;
+  Watch: (callback: WatchAction<T>, watchCfg?: WatchCfg) => WatcherType;
 };
 
 export type K8sUnfilteredActions<K extends KubernetesObject> = {
@@ -245,11 +231,6 @@ export type K8sInit<T extends GenericClass, K extends KubernetesObject> = K8sWit
      */
     InNamespace: (namespace: string) => K8sWithFilters<T, K>;
   };
-
-export type WatchAction<T extends GenericClass, K extends KubernetesObject = InstanceType<T>> = (
-  update: K,
-  phase: WatchPhase,
-) => Promise<void> | void;
 
 // Special types to handle the recursive keyof typescript lookup
 type Join<K, P> = K extends string | number
