+++ src/fluent/utils.ts
@@ -9,7 +9,8 @@ import { Agent as httpsAgent } from "https";
 import { fetch } from "../fetch";
 import { modelToGroupVersionKind } from "../kinds";
 import { GenericClass } from "../types";
-import { ApplyCfg, FetchMethods, Filters, K8sConfigPromise } from "./types";
+import { ApplyCfg, FetchMethods, K8sConfigPromise } from "./types";
+import { Filters } from "../types/shared";
 import fs from "fs";
 import { V1Eviction as Eviction } from "@kubernetes/client-node";
 const SSA_CONTENT_TYPE = "application/apply-patch+yaml";
